Make impl filename relative to current module.
The impl filename should be relative to the current module.

Since the Resolver already has the ast.Module instance, then the
filename is in the module's header.

If we re-resolve the module base filename, we may end up
elsewhere. However, the design says that the impls is relative to the
module base file. That is understood to be the case regardless of
redirection.