Don't require impls to be sorted.
I would like impls to be sorted and have the toolchain to
topologically sort the impl files.

However, this is quite a bit of work since it requires
https://clang.llvm.org/docs/StandardCPlusPlusModules.html#discovering-dependencies

Also, it's not clear yet how to achieve the Go-like module semantics
for C++ since on the one hand module partitions help with the problem,
but at the same time they can't have circular imports.