Simplify svars to not hold values and errors.
Since actions are using separate svars for values and errors, there's
no need to have an svar handle both.

This also disambiguates the case of getting a value from an
svar[error], and getting an error. Is it a stored error? It is a
cancellation? Is it a conversion error? Too many errors.